+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Mark opt in option `swiftAsyncStacktraces` as stable (#8373)
 - Add `dataCollection` option under `experimental` for configuring data scrubbing behavior (#8369)
+- Add scope API to clear feature flags (#8364)
 
 ## 9.21.0